examples/kubernetes-ingress/scripts: factor out cri-o installation into function
Since cri-o is now installed in the base image used in the Vagrantfile, do not
try to install it when provisioning the K8s VMs. Factor out this logic into
a separate function instead of removing it in case it is needed in the future.

Signed-off by: Ian Vernon <ian@cilium.io>